+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/prioritization/JobDataExtractor.java
@@ -26,13 +26,13 @@ import java.util.stream.Collectors;
 import javax.xml.bind.JAXBElement;
 import javax.xml.namespace.QName;
 
-import org.planqk.nisq.analyzer.core.connector.qiskit.QiskitSdkConnector;
 import org.planqk.nisq.analyzer.core.model.AnalysisJob;
 import org.planqk.nisq.analyzer.core.model.CircuitResult;
 import org.planqk.nisq.analyzer.core.model.CompilationJob;
 import org.planqk.nisq.analyzer.core.model.JobType;
 import org.planqk.nisq.analyzer.core.model.McdaJob;
 import org.planqk.nisq.analyzer.core.model.QpuSelectionJob;
+import org.planqk.nisq.analyzer.core.qprov.QProvService;
 import org.planqk.nisq.analyzer.core.repository.AnalysisJobRepository;
 import org.planqk.nisq.analyzer.core.repository.CompilationJobRepository;
 import org.planqk.nisq.analyzer.core.repository.McdaJobRepository;
@@ -73,7 +73,7 @@ public class JobDataExtractor {
 
     private final XmcdaRepository xmcdaRepository;
 
-    private final QiskitSdkConnector qiskitSdkConnector;
+    private final QProvService qProvService;
 
     /**
      * Get the required information to run MCDA methods from different kinds of NISQ Analyzer jobs
@@ -142,7 +142,7 @@ public class JobDataExtractor {
 
         for (CircuitResult result : circuitResults) {
 
-            int backendQueueSize = qiskitSdkConnector.getQueueSizeOfQpu(result.getQpu());
+            int backendQueueSize = qProvService.getQueueSizeOfQpu(result.getQpu());
 
             // add alternative representing the analysis result
             String name = result.getQpu() + "-" + result.getCompiler() + "-" + result.getCircuitName();
