+++ org.planqk.nisq.analyzer.core/src/main/resources/application-docker.properties
@@ -17,7 +17,11 @@ springdoc.swagger-ui.operationsSorter=alpha
 spring.datasource.initialization-mode=always
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 springdoc.default-produces-media-type=application/hal+json
-
+#----------------------------
+# QProv Service
+#----------------------------
+org.planqk.nisq.analyzer.qprov.hostname=localhost
+org.planqk.nisq.analyzer.qprov.port=5020
 #----------------------------
 # SDK Connector configuration
 #----------------------------
@@ -25,7 +29,6 @@ org.planqk.nisq.analyzer.connector.qiskit.hostname=${QISKIT_HOSTNAME}
 org.planqk.nisq.analyzer.connector.qiskit.port=${QISKIT_PORT}
 org.planqk.nisq.analyzer.connector.qiskit.version=v1.0
 org.planqk.nisq.analyzer.connector.qiskit.pollInterval=10000
-
 org.planqk.nisq.analyzer.connector.pytket.hostname=${PYTKET_HOSTNAME}
 org.planqk.nisq.analyzer.connector.pytket.port=${PYTKET_PORT}
 org.planqk.nisq.analyzer.connector.pytket.version=v1.0
