+++ org.planqk.nisq.analyzer.core/src/test/resources/application.properties
@@ -19,7 +19,11 @@ springdoc.swagger-ui.operationsSorter=alpha
 spring.datasource.initialization-mode=always
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 springdoc.default-produces-media-type=application/hal+json
-
+#----------------------------
+# QProv Service
+#----------------------------
+org.planqk.nisq.analyzer.qprov.hostname=localhost
+org.planqk.nisq.analyzer.qprov.port=5020
 #----------------------------
 # SDK Connector configuration
 #----------------------------
@@ -27,7 +31,6 @@ org.planqk.nisq.analyzer.connector.qiskit.hostname=localhost
 org.planqk.nisq.analyzer.connector.qiskit.port=5013
 org.planqk.nisq.analyzer.connector.qiskit.version=v1.0
 org.planqk.nisq.analyzer.connector.qiskit.pollInterval=10000
-
 org.planqk.nisq.analyzer.connector.pytket.hostname=localhost
 org.planqk.nisq.analyzer.connector.pytket.port=5015
 org.planqk.nisq.analyzer.connector.pytket.version=v1.0
